move coltype back
