+++ packages/fuel-indexer-types/Cargo.toml
@@ -18,5 +18,4 @@ fuel-types = { workspace = true }
 fuels = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
-strum = { version = "0.24", default-features = false, features = ["derive"] }
 tai64 = { version = "4.0", default-features = false, features = ["serde"] }
