--- packages/fuel-indexer-types/src/db.rs
@@ -1,188 +1,0 @@
-use std::fmt;
-use strum::AsRefStr;
-
-/// SQL database types used by indexers.
-#[derive(Copy, Clone, Debug, Eq, PartialEq, Default, AsRefStr)]
-pub enum ColumnType {
-    #[default]
-    ID = 0,
-    Address = 1,
-    AssetId = 2,
-    Bytes4 = 3,
-    Bytes8 = 4,
-    Bytes32 = 5,
-    ContractId = 6,
-    Salt = 7,
-    Int4 = 8,
-    Int8 = 9,
-    UInt4 = 10,
-    UInt8 = 11,
-    Timestamp = 12,
-    Blob = 13,
-    ForeignKey = 14,
-    Json = 15,
-    MessageId = 16,
-    Charfield = 17,
-    Identity = 18,
-    Boolean = 19,
-    Object = 20,
-    UInt16 = 21,
-    Int16 = 22,
-    Bytes64 = 23,
-    Signature = 24,
-    Nonce = 25,
-    HexString = 26,
-    Tai64Timestamp = 27,
-    TxId = 28,
-    BlockHeight = 29,
-    Enum = 30,
-    Int1 = 31,
-    UInt1 = 32,
-    Virtual = 33,
-    BlockId = 34,
-    Array = 35,
-}
-
-impl From<ColumnType> for i32 {
-    fn from(typ: ColumnType) -> i32 {
-        match typ {
-            ColumnType::ID => 0,
-            ColumnType::Address => 1,
-            ColumnType::AssetId => 2,
-            ColumnType::Bytes4 => 3,
-            ColumnType::Bytes8 => 4,
-            ColumnType::Bytes32 => 5,
-            ColumnType::ContractId => 6,
-            ColumnType::Salt => 7,
-            ColumnType::Int4 => 8,
-            ColumnType::Int8 => 9,
-            ColumnType::UInt4 => 10,
-            ColumnType::UInt8 => 11,
-            ColumnType::Timestamp => 12,
-            ColumnType::Blob => 13,
-            ColumnType::ForeignKey => 14,
-            ColumnType::Json => 15,
-            ColumnType::MessageId => 16,
-            ColumnType::Charfield => 17,
-            ColumnType::Identity => 18,
-            ColumnType::Boolean => 19,
-            ColumnType::Object => 20,
-            ColumnType::UInt16 => 21,
-            ColumnType::Int16 => 22,
-            ColumnType::Bytes64 => 23,
-            ColumnType::Signature => 24,
-            ColumnType::Nonce => 25,
-            ColumnType::HexString => 26,
-            ColumnType::Tai64Timestamp => 27,
-            ColumnType::TxId => 28,
-            ColumnType::BlockHeight => 29,
-            ColumnType::Enum => 30,
-            ColumnType::Int1 => 31,
-            ColumnType::UInt1 => 32,
-            ColumnType::Virtual => 33,
-            ColumnType::BlockId => 34,
-            ColumnType::Array => 35,
-        }
-    }
-}
-
-impl From<ColumnType> for i64 {
-    fn from(typ: ColumnType) -> i64 {
-        let typ = i32::from(typ);
-        typ as i64
-    }
-}
-
-impl fmt::Display for ColumnType {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{self:?}")
-    }
-}
-
-impl From<i32> for ColumnType {
-    fn from(num: i32) -> ColumnType {
-        match num {
-            0 => ColumnType::ID,
-            1 => ColumnType::Address,
-            2 => ColumnType::AssetId,
-            3 => ColumnType::Bytes4,
-            4 => ColumnType::Bytes8,
-            5 => ColumnType::Bytes32,
-            6 => ColumnType::ContractId,
-            7 => ColumnType::Salt,
-            8 => ColumnType::Int4,
-            9 => ColumnType::Int8,
-            10 => ColumnType::UInt4,
-            11 => ColumnType::UInt8,
-            12 => ColumnType::Timestamp,
-            13 => ColumnType::Blob,
-            14 => ColumnType::ForeignKey,
-            15 => ColumnType::Json,
-            16 => ColumnType::MessageId,
-            17 => ColumnType::Charfield,
-            18 => ColumnType::Identity,
-            19 => ColumnType::Boolean,
-            20 => ColumnType::Object,
-            21 => ColumnType::Int16,
-            22 => ColumnType::UInt16,
-            23 => ColumnType::Bytes64,
-            24 => ColumnType::Signature,
-            25 => ColumnType::Nonce,
-            26 => ColumnType::HexString,
-            27 => ColumnType::Tai64Timestamp,
-            28 => ColumnType::TxId,
-            29 => ColumnType::BlockHeight,
-            30 => ColumnType::Enum,
-            31 => ColumnType::Int1,
-            32 => ColumnType::UInt1,
-            33 => ColumnType::Virtual,
-            34 => ColumnType::BlockId,
-            35 => ColumnType::Array,
-            _ => unimplemented!("Invalid ColumnType: {num}."),
-        }
-    }
-}
-
-impl From<&str> for ColumnType {
-    fn from(name: &str) -> ColumnType {
-        match name {
-            "ID" => ColumnType::ID,
-            "Address" => ColumnType::Address,
-            "AssetId" => ColumnType::AssetId,
-            "Bytes4" => ColumnType::Bytes4,
-            "Bytes8" => ColumnType::Bytes8,
-            "Bytes32" => ColumnType::Bytes32,
-            "ContractId" => ColumnType::ContractId,
-            "Salt" => ColumnType::Salt,
-            "Int4" => ColumnType::Int4,
-            "Int8" => ColumnType::Int8,
-            "UInt4" => ColumnType::UInt4,
-            "UInt8" => ColumnType::UInt8,
-            "Timestamp" => ColumnType::Timestamp,
-            "Blob" => ColumnType::Blob,
-            "ForeignKey" => ColumnType::ForeignKey,
-            "Json" => ColumnType::Json,
-            "MessageId" => ColumnType::MessageId,
-            "Charfield" => ColumnType::Charfield,
-            "Identity" => ColumnType::Identity,
-            "Boolean" => ColumnType::Boolean,
-            "Object" => ColumnType::Object,
-            "UInt16" => ColumnType::UInt16,
-            "Int16" => ColumnType::Int16,
-            "Bytes64" => ColumnType::Bytes64,
-            "Signature" => ColumnType::Signature,
-            "Nonce" => ColumnType::Nonce,
-            "HexString" => ColumnType::HexString,
-            "Tai64Timestamp" => ColumnType::Tai64Timestamp,
-            "TxId" => ColumnType::TxId,
-            "BlockHeight" => ColumnType::BlockHeight,
-            "Enum" => ColumnType::Enum,
-            "Int1" => ColumnType::Int1,
-            "UInt1" => ColumnType::UInt1,
-            "Virtual" => ColumnType::Virtual,
-            "BlockId" => ColumnType::BlockId,
-            "Array" => ColumnType::Array,
-            _ => unimplemented!("Invalid ColumnType: '{name}'."),
-        }
-    }
-}
